requirements-webhook.txt
# Requirements for webhook service
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
livekit-api>=1.0.0
httpx>=0.25.0
